requirements.txt
pyamg
matplotlib
vedo
pyinstrument
pyyaml
scikit-fem
